extract a util
